refactor to remove the utils subdirectory
