+++ rfmux/core/transferfunctions.py
